[deform_conv] CB-6802: Fix pass split_deformable_conv
gen_offset_base from DeformConvOp returns int64 by default.
Change its dtype to be same as offset input in split_deformable_conv.

+++ AIPUBuilder/Parser/ops/onnx_ops/nn_ops.py
@@ -492,7 +492,7 @@ class DeformConvOp(BaseConvOp, OnnxOp):
         kernel_pos_w, kernel_pos_h = np.meshgrid(np.arange(0, kw_new, dw), np.arange(0, kh_new, dh))
         kernel_pos = np.stack([kernel_pos_h, kernel_pos_w], axis=2)  # shape: [kh, kw, 2]
 
-        kernel_offset = np.zeros([oh, ow, kh, kw, 2])
+        kernel_offset = np.zeros([oh, ow, kh, kw, 2], dtype=np.int64)
         for i in range(oh):
             h_coord = bh + sth * i
             for j in range(ow):
